services/backend-for-frontend/requirements.txt
Flask
flask-cors
requests
gunicorn
python-dotenv
